+++ src/API/Type/FindFolderParentType.php
@@ -24,7 +24,7 @@ use jamesiarmes\PEWS\API\Type;
  * @method ArrayOfFoldersType getFolders()
  * @method FindFolderParentType setFolders(ArrayOfFoldersType $folders)
  */
-class FindFolderParentType extends Type
+class FindFolderParentType extends Type implements \Countable, \ArrayAccess, \IteratorAggregate
 {
 
     /**
@@ -56,4 +56,38 @@ class FindFolderParentType extends Type
      * @var \jamesiarmes\PEWS\API\Type\ArrayOfFoldersType
      */
     protected $folders = null;
+
+    public function count()
+    {
+        return count($this->folders);
+    }
+
+    public function offsetExists($offset)
+    {
+        return isset($this->folders[$offset]);
+    }
+
+    public function offsetGet($offset)
+    {
+        return isset($this->folders[$offset]);
+    }
+
+    public function offsetSet($offset, $value)
+    {
+        if (is_null($offset)) {
+            array_push($this->folders, $value);
+        } else {
+            $this->folders[$offset] = $value;
+        }
+    }
+
+    public function offsetUnset($offset)
+    {
+        unset($this->folders[$offset]);
+    }
+
+    public function getIterator()
+    {
+        return new \ArrayIterator($this->folders->getIterator());
+    }
 }